chore: reconstructing config as monolith
