+++ .storybook/manager.js
@@ -1,4 +1,4 @@
-import { addons } from '@storybook/addons';
+import { addons } from '@storybook/manager-api';
 import autoGuruTheme from './ag-theme';
 
 addons.setConfig({
