--- .storybook/vanilla-extract.js
@@ -1,39 +1,0 @@
-const { VanillaExtractPlugin } = require('@vanilla-extract/webpack-plugin');
-const MiniCssExtractPlugin = require('mini-css-extract-plugin');
-
-module.exports = {
-	webpackFinal(baseConfig, options) {
-		const { module = {}, plugins = {} } = baseConfig;
-
-		return {
-			...baseConfig,
-			plugins: [
-				...plugins,
-				new VanillaExtractPlugin(),
-				new MiniCssExtractPlugin(),
-			],
-			module: {
-				...module,
-				rules: [
-					...((module.rules &&
-						module.rules.filter(
-							(rule) => !rule?.test?.test('test.css'),
-						)) ||
-						[]),
-					{
-						test: /\.vanilla\.css$/i, // Targets only CSS files generated by vanilla-extract
-						use: [
-							MiniCssExtractPlugin.loader,
-							{
-								loader: require.resolve('css-loader'),
-								options: {
-									url: true, // Required as image imports should be handled via JS/TS import statements
-								},
-							},
-						],
-					},
-				],
-			},
-		};
-	},
-};
