--- packages/overdrive/.storybook/ag-theme.js
@@ -1,14 +1,0 @@
-import { create } from '@storybook/theming';
-
-export default create({
-	base: 'dark',
-	brandTitle: 'AutoGuru SDK Client Components',
-	brandUrl: 'https://www.autoguru.com.au/',
-	brandImage: 'https://cdn.autoguru.com.au/images/logos/otto-white.svg',
-	brandTarget: '_self',
-	colorPrimary: '#ffc001',
-	colorSecondary: '#03af83',
-	textInverseColor: '#ffc001',
-	barSelectedColor: '#ffc001',
-	appBorderRadius: 3,
-});
