--- packages/overdrive/.storybook/manager.js
@@ -1,6 +1,0 @@
-import { addons } from '@storybook/manager-api';
-import autoGuruTheme from './ag-theme';
-
-addons.setConfig({
-	theme: autoGuruTheme,
-});
