--- packages/overdrive/package.json
@@ -1,97 +1,0 @@
-{
-  "name": "@autoguru/overdrive",
-  "version": "4.34.0",
-  "description": "Overdrive is a product component library, and design system for AutoGuru.",
-  "types": "dist/index.d.ts",
-  "main": "dist/index.js",
-  "keywords": [
-    "component",
-    "design-system",
-    "react",
-    "storybook",
-    "styleguide",
-    "ui",
-    "autoguru"
-  ],
-  "homepage": "https://overdrive.autoguru.io",
-  "repository": "autoguru-au/overdrive",
-  "license": "MIT",
-  "sideEffects": [
-    "./lib/reset/**/*",
-    "./lib/theme/**/*",
-    "./lib/themes/**/*",
-    "./dist/reset/**/*",
-    "./dist/theme/**/*",
-    "./dist/themes/**/*"
-  ],
-  "files": [
-    "dist"
-  ],
-  "publishConfig": {
-    "access": "public"
-  },
-  "scripts": {
-    "lint": "tsc --noEmit --skipLibCheck",
-    "typeEmit": "tsc -d --declarationDir dist --emitDeclarationOnly",
-    "build": "babel lib --out-dir dist --extensions '.ts,.tsx, .css'",
-    "prepublishOnly": "yarn run build && yarn run typeEmit",
-    "watch": "tsc --watch",
-    "storybook": "storybook dev -p 6006",
-    "build-storybook": "storybook build"
-  },
-  "devDependencies": {
-    "@autoguru/icons": "^1.7.28",
-    "@autoguru/tsconfig": "1.0.80",
-    "@autoguru/utilities": "1.1.1",
-    "@chromatic-com/storybook": "^3.2.2",
-    "@popperjs/core": "2.11.6",
-    "@react-stately/toggle": "3.7.7",
-    "@storybook/addon-a11y": "^8.4.4",
-    "@storybook/addon-essentials": "^8.4.4",
-    "@storybook/addon-links": "^8.4.4",
-    "@storybook/addon-interactions": "^8.4.4",
-    "@storybook/addon-onboarding": "^8.4.4",
-    "@storybook/blocks": "^8.4.4",
-    "@storybook/react": "^8.4.4",
-    "@storybook/react-vite": "^8.4.4",
-    "@storybook/test": "^8.4.4",
-    "@types/react": "^18.0.17",
-    "@types/react-dom": "^18.0.6",
-    "@vanilla-extract/dynamic": "^2.1.2",
-    "@vanilla-extract/vite-plugin": "^4.0.17",
-    "clsx": "1.2.1",
-    "colord": "2.9.3",
-    "csstype": "^3.1.3",
-    "deepmerge": "^4.2.2",
-    "eslint-plugin-jest": "26.9.0",
-    "intersection-observer": "^0.12.2",
-    "prop-types": "^15.8.1",
-    "react": "^18.2.0",
-    "react-aria": "3.34.2",
-    "react-dom": "^18.2.0",
-    "react-focus-lock": "2.9.4",
-    "react-intersection-observer": "^9.4.3",
-    "react-keyed-flatten-children": "1.3.0",
-    "react-swipeable": "^7.0.0",
-    "react-types": "0.1.0",
-    "storybook": "^8.4.4",
-    "webpack": "5.77.0"
-  },
-  "peerDependencies": {
-    "@autoguru/icons": "^1.7.28",
-    "@autoguru/utilities": "^1.0.98",
-    "@popperjs/core": "^2.10.2",
-    "@vanilla-extract/dynamic": "^2.1.2",
-    "clsx": "^1.1.1",
-    "colord": "^2.0.1",
-    "csstype": "^3.1.3",
-    "react": ">=16.8",
-    "react-aria": "3.34.2",
-    "react-dom": ">=16.8",
-    "react-focus-lock": "^2.5.0",
-    "react-intersection-observer": ">=9.4.3",
-    "react-keyed-flatten-children": "^1.3.0",
-    "react-swipeable": ">=7.0.0",
-    "webpack": "*"
-  }
-}
