--- readme.md
@@ -1,1 +1,0 @@
-packages/overdrive/readme.md
