+++ scripts/babel.config.js
